backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
alembic==1.16.4
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.3
python-multipart==0.0.20
redis==6.4.0
celery==5.5.3
elasticsearch==8.15.1
httpx==0.28.1
email-validator==2.2.0
python-dateutil==2.9.0.post0
openpyxl==3.1.5
pandas==2.2.3
pytest==8.4.1
pytest-asyncio==1.1.0
